requirements.txt
langchain
langchain-openai
streamlit
datasets
chromadb
python-dotenv
